Girls that are taken off work by the matron due to tiredness will now have their jobs listed as "[z] " followed by the stored job title (example "[z] Security"), instead of just being listed as "Free Time". This way, it can be seen at a glance that she's just temporarily off work due to tiredness rather than just doing nothing, and you can see what job she will be going back to when the matron puts her back to work.
I moved code for cBrothel->is_she_resting(sGirl *girl) up to cGirls, so it's more widely accessible. For simplicity it's currently just redirecting calls for cBrothel->is_she_resting() to the new method, but it would probably be better to change all references to point to the new location in cGirl->is_resting() and remove the old one in cBrothel->is_she_resting().