Update WorkBarSinger.cpp
Changed the layout a bit. I was confused by the jobperformance part so i shortened it a little bit (didn't remove anything just removed some tabs so you get a clearer view of the structure.

Then i added the I_confidence parameter in the beginning of the document because there was a reward with confidence in it. rewrote that parte so it added a point to the parameter and rewrote the improve stats part to include all the major gains necessary for the report.

If i forgot anything pls respond, other wise i'll just continue ;)